Cardano.Ledger.Babbage.PParams

Description

This module contains the type of protocol parameters and EraPParams instance

Documentation

class AlonzoEraPParams era ⇒ BabbageEraPParams era where

Methods

hkdCoinsPerUTxOByteL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f CoinPerByte)

newtype CoinPerByte

Constructors

CoinPerByte

Fields

unCoinPerByte ∷ Coin

ppCoinsPerUTxOByteL ∷ ∀ era. BabbageEraPParams era ⇒ Lens' (PParams era) CoinPerByte

ppuCoinsPerUTxOByteL ∷ ∀ era. BabbageEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe CoinPerByte)

data BabbagePParams f era

Babbage Protocol parameters. Ways in which parameters have changed from Alonzo: lack of d, extraEntropy and replacement of coinsPerUTxOWord with coinsPerUTxOByte

Constructors

BabbagePParams

Fields

bppMinFeeA ∷ !(HKD f Coin)

The linear factor for the minimum fee calculation

bppMinFeeB ∷ !(HKD f Coin)

The constant factor for the minimum fee calculation

bppMaxBBSize ∷ !(HKD f Word32)

Maximal block body size

bppMaxTxSize ∷ !(HKD f Word32)

Maximal transaction size

bppMaxBHSize ∷ !(HKD f Word16)

Maximal block header size

bppKeyDeposit ∷ !(HKD f Coin)

The amount of a key registration deposit

bppPoolDeposit ∷ !(HKD f Coin)

The amount of a pool registration deposit

bppEMax ∷ !(HKD f EpochInterval)

Maximum number of epochs in the future a pool retirement is allowed to be scheduled for.

bppNOpt ∷ !(HKD f Word16)

Desired number of pools

bppA0 ∷ !(HKD f NonNegativeInterval)

Pool influence

bppRho ∷ !(HKD f UnitInterval)

Monetary expansion

bppTau ∷ !(HKD f UnitInterval)

Treasury expansion

bppProtocolVersion ∷ !(HKD f ProtVer)

Protocol version

bppMinPoolCost ∷ !(HKD f Coin)

Minimum Stake Pool Cost

bppCoinsPerUTxOByte ∷ !(HKD f CoinPerByte)

Cost in lovelace per byte of UTxO storage (instead of bppCoinsPerUTxOByte)

bppCostModels ∷ !(HKD f CostModels)

Cost models for non-native script languages

bppPrices ∷ !(HKD f Prices)

Prices of execution units (for non-native script languages)

bppMaxTxExUnits ∷ !(HKD f OrdExUnits)

Max total script execution resources units allowed per tx

bppMaxBlockExUnits ∷ !(HKD f OrdExUnits)

Max total script execution resources units allowed per block

bppMaxValSize ∷ !(HKD f Natural)

Max size of a Value in an output

bppCollateralPercentage ∷ !(HKD f Natural)

Percentage of the txfee which must be provided as collateral when including non-native scripts.

bppMaxCollateralInputs ∷ !(HKD f Natural)

Maximum number of collateral inputs allowed in a transaction

emptyBabbagePParams ∷ ∀ era. Era era ⇒ BabbagePParams Identity era

Returns a basic "empty" PParams structure with all zero values.

emptyBabbagePParamsUpdate ∷ BabbagePParams StrictMaybe era

data DowngradeBabbagePParams f

Constructors

DowngradeBabbagePParams

Fields

dbppD ∷ !(HKD f UnitInterval)

dbppExtraEntropy ∷ !(HKD f Nonce)

upgradeBabbagePParams ∷ ∀ f. HKDFunctor f ⇒ Bool → PParamsHKD f AlonzoEra → BabbagePParams f BabbageEra

getLanguageView ∷ AlonzoEraPParams era ⇒ PParams era → Language → LangDepView

data LangDepView

Constructors

LangDepView

Fields

tag ∷ ByteString

params ∷ ByteString

encodeLangViews ∷ Set LangDepView → Encoding

coinsPerUTxOWordToCoinsPerUTxOByte ∷ CoinPerWord → CoinPerByte

A word is 8 bytes, so convert from coinsPerUTxOWord to coinsPerUTxOByte, rounding down.

coinsPerUTxOByteToCoinsPerUTxOWord ∷ CoinPerByte → CoinPerWord

A word is 8 bytes, so convert from coinsPerUTxOByte to coinsPerUTxOWord.

babbagePParamsHKDPairs ∷ ∀ era f. (BabbageEraPParams era, HKDFunctor f, ProtVerAtMost era 8) ⇒ Proxy f → PParamsHKD f era → [(Key, HKD f Value)]

babbageCommonPParamsHKDPairs ∷ ∀ era f. (BabbageEraPParams era, HKDFunctor f) ⇒ Proxy f → PParamsHKD f era → [(Key, HKD f Value)]

These are the fields that are common across all eras starting with Babbage.

Orphan instances

AlonzoEraPParams BabbageEra

Instance details

Methods

hkdCoinsPerUTxOWordL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, ExactEra AlonzoEra BabbageEra) ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f CoinPerWord)

hkdCostModelsL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f CostModels)

hkdPricesL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f Prices)

hkdMaxTxExUnitsL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f ExUnits)

hkdMaxBlockExUnitsL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f ExUnits)

hkdMaxValSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f Natural)

hkdCollateralPercentageL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f Natural)

hkdMaxCollateralInputsL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f Natural)

EraPParams BabbageEra

Instance details

Associated Types

type PParamsHKD f BabbageEra = (r ∷ Type)

type UpgradePParams f BabbageEra

type DowngradePParams f BabbageEra

Methods

applyPPUpdates ∷ PParams BabbageEra → PParamsUpdate BabbageEra → PParams BabbageEra

emptyPParamsIdentity ∷ PParamsHKD Identity BabbageEra

emptyPParamsStrictMaybe ∷ PParamsHKD StrictMaybe BabbageEra

upgradePParamsHKD ∷ ∀ (f ∷ Type → Type). (HKDApplicative f, EraPParams (PreviousEra BabbageEra)) ⇒ UpgradePParams f BabbageEra → PParamsHKD f (PreviousEra BabbageEra) → PParamsHKD f BabbageEra

downgradePParamsHKD ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, EraPParams (PreviousEra BabbageEra)) ⇒ DowngradePParams f BabbageEra → PParamsHKD f BabbageEra → PParamsHKD f (PreviousEra BabbageEra)

hkdMinFeeAL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f Coin)

hkdMinFeeBL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f Coin)

hkdMaxBBSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f Word32)

hkdMaxTxSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f Word32)

hkdMaxBHSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f Word16)

hkdKeyDepositL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f Coin)

hkdPoolDepositL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f Coin)

hkdEMaxL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f EpochInterval)

hkdNOptL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f Word16)

hkdA0L ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f NonNegativeInterval)

hkdRhoL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f UnitInterval)

hkdTauL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f UnitInterval)

hkdDL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, ProtVerAtMost BabbageEra 6) ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f UnitInterval)

ppDG ∷ SimpleGetter (PParams BabbageEra) UnitInterval

hkdExtraEntropyL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, ProtVerAtMost BabbageEra 6) ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f Nonce)

hkdProtocolVersionL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, ProtVerAtMost BabbageEra 8) ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f ProtVer)

ppProtocolVersionL ∷ Lens' (PParams BabbageEra) ProtVer

ppuProtocolVersionL ∷ Lens' (PParamsUpdate BabbageEra) (StrictMaybe ProtVer)

hkdMinUTxOValueL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, ProtVerAtMost BabbageEra 4) ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f Coin)

hkdMinPoolCostL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f BabbageEra) (HKD f Coin)

EraGov BabbageEra

Instance details

Associated Types

type GovState BabbageEra = (r ∷ Type)

Methods

emptyGovState ∷ GovState BabbageEra

curPParamsGovStateL ∷ Lens' (GovState BabbageEra) (PParams BabbageEra)

prevPParamsGovStateL ∷ Lens' (GovState BabbageEra) (PParams BabbageEra)

futurePParamsGovStateL ∷ Lens' (GovState BabbageEra) (FuturePParams BabbageEra)

obligationGovState ∷ GovState BabbageEra → Obligations