Cardano.Ledger.Babbage.Rules.Utxo

Documentation

data BabbageUTXO era

data BabbageUtxoPredFailure era

Predicate failure for the Babbage Era

Constructors

AlonzoInBabbageUtxoPredFailure !(AlonzoUtxoPredFailure era)

IncorrectTotalCollateralField
The collateral is not equivalent to the total collateral asserted by the transaction

Fields

!Coin

collateral provided

!Coin

collateral amount declared in transaction body

BabbageOutputTooSmallUTxO ![(TxOut era, Coin)]
list of supplied transaction outputs that are too small, together with the minimum value for the given output.

BabbageNonDisjointRefInputs !(NonEmpty (TxIn (EraCrypto era)))
TxIns that appear in both inputs and reference inputs

utxoTransition ∷ ∀ era. (EraUTxO era, BabbageEraTxBody era, AlonzoEraTxWits era, Tx era ~ AlonzoTx era, EraRule "UTXO" era ~ BabbageUTXO era, InjectRuleFailure "UTXO" ShelleyUtxoPredFailure era, InjectRuleFailure "UTXO" AllegraUtxoPredFailure era, InjectRuleFailure "UTXO" AlonzoUtxoPredFailure era, InjectRuleFailure "UTXO" BabbageUtxoPredFailure era, Embed (EraRule "UTXOS" era) (BabbageUTXO era), Environment (EraRule "UTXOS" era) ~ UtxoEnv era, State (EraRule "UTXOS" era) ~ UTxOState era, Signal (EraRule "UTXOS" era) ~ Tx era) ⇒ TransitionRule (EraRule "UTXO" era)

The UTxO transition rule for the Babbage eras.

feesOK ∷ ∀ era rule. (EraUTxO era, BabbageEraTxBody era, AlonzoEraTxWits era, InjectRuleFailure rule AlonzoUtxoPredFailure era, InjectRuleFailure rule BabbageUtxoPredFailure era) ⇒ PParams era → Tx era → UTxO era → Test (EraRuleFailure rule era)

feesOK is a predicate with several parts. Some parts only apply in special circumstances. 1) The fee paid is >= the minimum fee 2) If the total ExUnits are 0 in both Memory and Steps, no further part needs to be checked. 3) The collateral consists only of VKey addresses 4) The collateral inputs do not contain any non-ADA part 5) The collateral is sufficient to cover the appropriate percentage of the fee marked in the transaction 6) The collateral is equivalent to total collateral asserted by the transaction 7) There is at least one collateral input

feesOK can differ from Era to Era, as new notions of fees arise. This is the Babbage version See: Figure 2: Functions related to fees and collateral, in the Babbage specification In the spec feesOK is a boolean function. Because wee need to handle predicate failures in the implementaion, it is coded as a Test. Which is a validation. This version is generic in that it can be lifted to any PredicateFailure type that embeds BabbageUtxoPred era. This makes it possibly useful in future Eras.

validateTotalCollateral ∷ ∀ era rule. (BabbageEraTxBody era, InjectRuleFailure rule AlonzoUtxoPredFailure era, InjectRuleFailure rule BabbageUtxoPredFailure era) ⇒ PParams era → TxBody era → Map (TxIn (EraCrypto era)) (TxOut era) → Test (EraRuleFailure rule era)

validateCollateralEqBalance ∷ Coin → StrictMaybe Coin → Validation (NonEmpty (BabbageUtxoPredFailure era)) ()

validateOutputTooSmallUTxO ∷ (EraTxOut era, Foldable f) ⇒ PParams era → f (Sized (TxOut era)) → Test (BabbageUtxoPredFailure era)

Orphan instances

InjectRuleFailure "UTXO" AllegraUtxoPredFailure (BabbageEra c)

Instance details

Methods

injectFailure ∷ AllegraUtxoPredFailure (BabbageEra c) → EraRuleFailure "UTXO" (BabbageEra c)

InjectRuleFailure "UTXO" AlonzoUtxoPredFailure (BabbageEra c)

Instance details

Methods

injectFailure ∷ AlonzoUtxoPredFailure (BabbageEra c) → EraRuleFailure "UTXO" (BabbageEra c)

InjectRuleFailure "UTXO" AlonzoUtxosPredFailure (BabbageEra c)

Instance details

Methods

injectFailure ∷ AlonzoUtxosPredFailure (BabbageEra c) → EraRuleFailure "UTXO" (BabbageEra c)

InjectRuleFailure "UTXO" ShelleyPpupPredFailure (BabbageEra c)

Instance details

Methods

injectFailure ∷ ShelleyPpupPredFailure (BabbageEra c) → EraRuleFailure "UTXO" (BabbageEra c)

InjectRuleFailure "UTXO" ShelleyUtxoPredFailure (BabbageEra c)

Instance details

Methods

injectFailure ∷ ShelleyUtxoPredFailure (BabbageEra c) → EraRuleFailure "UTXO" (BabbageEra c)

(EraTx era, EraUTxO era, BabbageEraTxBody era, AlonzoEraTxWits era, Tx era ~ AlonzoTx era, EraRule "UTXO" era ~ BabbageUTXO era, InjectRuleFailure "UTXO" ShelleyUtxoPredFailure era, InjectRuleFailure "UTXO" AllegraUtxoPredFailure era, InjectRuleFailure "UTXO" AlonzoUtxoPredFailure era, InjectRuleFailure "UTXO" BabbageUtxoPredFailure era, Embed (EraRule "UTXOS" era) (BabbageUTXO era), Environment (EraRule "UTXOS" era) ~ UtxoEnv era, State (EraRule "UTXOS" era) ~ UTxOState era, Signal (EraRule "UTXOS" era) ~ Tx era, PredicateFailure (EraRule "UTXO" era) ~ BabbageUtxoPredFailure era) ⇒ STS (BabbageUTXO era)

Instance details

Associated Types

type State (BabbageUTXO era)

type Signal (BabbageUTXO era)

type Environment (BabbageUTXO era)

type BaseM (BabbageUTXO era) ∷ Type → Type

type Event (BabbageUTXO era)

type PredicateFailure (BabbageUTXO era)

Methods

initialRules ∷ [InitialRule (BabbageUTXO era)]

transitionRules ∷ [TransitionRule (BabbageUTXO era)]

assertions ∷ [Assertion (BabbageUTXO era)]

renderAssertionViolation ∷ AssertionViolation (BabbageUTXO era) → String

(Era era, STS (BabbageUTXOS era), PredicateFailure (EraRule "UTXOS" era) ~ AlonzoUtxosPredFailure era, Event (EraRule "UTXOS" era) ~ Event (BabbageUTXOS era)) ⇒ Embed (BabbageUTXOS era) (BabbageUTXO era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (BabbageUTXOS era) → PredicateFailure (BabbageUTXO era)

wrapEvent ∷ Event (BabbageUTXOS era) → Event (BabbageUTXO era)