Cardano.Ledger.Babbage.Tx

Documentation

alonzoInputHashes :: (HasField "inputs" (TxBody era) (Set (TxIn (Crypto era))), ValidateScript era, TxOut era ~ TxOut era) => Map (ScriptHash (Crypto era)) (Script era) -> ValidatedTx era -> UTxO era -> (Set (DataHash (Crypto era)), Set (TxIn (Crypto era)))

isTwoPhaseScriptAddressFromMap :: ValidateScript era => Map (ScriptHash (Crypto era)) (Script era) -> Addr (Crypto era) -> Bool

Compute if an Addr has the hash of a TwoPhaseScript, we can tell what kind of Script from the Hash, by looking it up in the Map

toCBORForMempoolSubmission :: (Typeable era, ToCBOR (TxBody era), ToCBOR (AuxiliaryData era)) => ValidatedTx era -> Encoding

Encode to CBOR for the purposes of transmission from node to node, or from wallet to node.

Note that this serialisation is neither the serialisation used on-chain (where Txs are deconstructed using segwit), nor the serialisation used for computing the transaction size (which omits the IsValid field for compatibility with Mary - see toCBORForSizeComputation).

segwitTx :: Annotator (TxBody era) -> Annotator (TxWitness era) -> IsValid -> Maybe (Annotator (AuxiliaryData era)) -> Annotator (ValidatedTx era)

indexedRdmrs :: (Era era, HasField "inputs" (TxBody era) (Set (TxIn (Crypto era))), HasField "wdrls" (TxBody era) (Wdrl (Crypto era)), HasField "certs" (TxBody era) (StrictSeq (DCert (Crypto era))), HasField "wits" tx (TxWitness era), HasField "body" tx (TxBody era)) => tx -> ScriptPurpose (Crypto era) -> Maybe (Data era, ExUnits)

Find the Data and ExUnits assigned to a script.

getMapFromValue :: Value crypto -> Map (PolicyID crypto) (Map AssetName Integer)

rdptrInv :: (HasField "inputs" (TxBody era) (Set (TxIn (Crypto era))), HasField "wdrls" (TxBody era) (Wdrl (Crypto era)), HasField "certs" (TxBody era) (StrictSeq (DCert (Crypto era))), HasField "minted" (TxBody era) (Set (ScriptHash (Crypto era)))) => TxBody era -> RdmrPtr -> StrictMaybe (ScriptPurpose (Crypto era))

rdptr :: (HasField "inputs" (TxBody era) (Set (TxIn (Crypto era))), HasField "wdrls" (TxBody era) (Wdrl (Crypto era)), HasField "certs" (TxBody era) (StrictSeq (DCert (Crypto era))), HasField "minted" (TxBody era) (Set (ScriptHash (Crypto era)))) => TxBody era -> ScriptPurpose (Crypto era) -> StrictMaybe RdmrPtr

totExUnits :: (HasField "wits" (Tx era) (Witnesses era), HasField "txrdmrs" (Witnesses era) (Redeemers era)) => Tx era -> ExUnits

minfee :: (HasField "_minfeeA" (PParams era) Natural, HasField "_minfeeB" (PParams era) Natural, HasField "_prices" (PParams era) Prices, HasField "wits" (Tx era) (Witnesses era), HasField "txrdmrs" (Witnesses era) (Redeemers era), HasField "txsize" (Tx era) Integer) => PParams era -> Tx era -> Coin

toCBORForSizeComputation :: (Typeable era, ToCBOR (TxBody era), ToCBOR (AuxiliaryData era)) => ValidatedTx era -> Encoding

This ensures that the size of transactions from Mary is unchanged. The individual components all store their bytes; the only work we do in this function is concatenating

isTwoPhaseScriptAddress :: ValidateScript era => ValidatedTx era -> Addr (Crypto era) -> Bool

hashScriptIntegrity :: Era era => Set LangDepView -> Redeemers era -> TxDats era -> StrictMaybe (ScriptIntegrityHash (Crypto era))

getCoin :: Era era => TxOut era -> Coin

newtype IsValid

Tag indicating whether non-native scripts in this transaction are expected to validate. This is added by the block creator when constructing the block.

Constructors

IsValid Bool

data ValidatedTx era

Constructors

ValidatedTx

Fields

body :: !(TxBody era)

wits :: !(TxWitness era)

isValid :: !IsValid

auxiliaryData :: !(StrictMaybe (AuxiliaryData era))

data ScriptIntegrity era

A ScriptIntegrityHash is the hash of three things. The first two come from the witnesses and the last comes from the Protocol Parameters.

Constructors

ScriptIntegrity !(Redeemers era) !(TxDats era) !(Set LangDepView)

data ScriptPurpose crypto

Constructors

Minting !(PolicyID crypto)

Spending !(TxIn crypto)

Rewarding !(RewardAcnt crypto)

Certifying !(DCert crypto)

class Indexable elem container where

Methods

indexOf :: elem -> container -> StrictMaybe Word64

fromIndex :: Word64 -> container -> StrictMaybe elem

txdats' :: TxWitness era -> TxDats era

txrdmrs :: TxWitness era -> (Era era, Script era ~ Script era) => Redeemers era

txscripts' :: TxWitness era -> Map (ScriptHash (Crypto era)) (Script era)

type ScriptIntegrityHash crypto = SafeHash crypto EraIndependentScriptIntegrity

hashData :: Era era => Data era -> DataHash (Crypto era)

data Data era

getLanguageView :: HasField "_costmdls" (PParams era) CostModels => PParams era -> Language -> LangDepView

data CostModel

A language dependent cost model for the Plutus evaluator. Note that the EvaluationContext is entirely dependent on the cost model parameters (ie the Map Text Integer) and that this type uses the smart constructor mkCostModel to hide the evaluation context.

nonNativeLanguages :: [Language]

txouts :: Era era => TxBody era -> UTxO era

Compute the transaction outputs of a transaction.

data EraIndependentScriptIntegrity

type DataHash crypto = SafeHash crypto EraIndependentData

data TxBody era where

Bundled Patterns

pattern TxBody :: BabbageBody era => Set (TxIn (Crypto era)) -> Set (TxIn (Crypto era)) -> Set (TxIn (Crypto era)) -> StrictSeq (Sized (TxOut era)) -> StrictMaybe (Sized (TxOut era)) -> StrictMaybe Coin -> StrictSeq (DCert (Crypto era)) -> Wdrl (Crypto era) -> Coin -> ValidityInterval -> StrictMaybe (Update era) -> Set (KeyHash 'Witness (Crypto era)) -> Value (Crypto era) -> StrictMaybe (ScriptIntegrityHash (Crypto era)) -> StrictMaybe (AuxiliaryDataHash (Crypto era)) -> StrictMaybe Network -> TxBody era