Cardano.Ledger.Babbage.TxBody

Documentation

data BabbageTxOut era

Constructors

TxOutCompactDatum !(CompactAddr (EraCrypto era)) !(CompactForm (Value era)) !(BinaryData era)

TxOutCompactRefScript !(CompactAddr (EraCrypto era)) !(CompactForm (Value era)) !(Datum era) !(Script era)

Bundled Patterns

pattern BabbageTxOut ∷ (Era era, Val (Value era), HasCallStack) ⇒ Addr (EraCrypto era) → Value era → Datum era → StrictMaybe (Script era) → BabbageTxOut era

pattern TxOutCompact ∷ (Era era, Val (Value era), Compactible (Value era), HasCallStack) ⇒ CompactAddr (EraCrypto era) → CompactForm (Value era) → BabbageTxOut era

pattern TxOutCompactDH ∷ (Era era, Val (Value era), Compactible (Value era), HasCallStack) ⇒ CompactAddr (EraCrypto era) → CompactForm (Value era) → DataHash (EraCrypto era) → BabbageTxOut era

allSizedOutputsBabbageTxBodyF ∷ BabbageEraTxBody era ⇒ SimpleGetter (TxBody era) (StrictSeq (Sized (TxOut era)))

babbageMinUTxOValue ∷ BabbageEraPParams era ⇒ PParams era → Sized a → Coin

data BabbageTxBody era where

Bundled Patterns

pattern BabbageTxBody ∷ BabbageEraTxBody era ⇒ Set (TxIn (EraCrypto era)) → Set (TxIn (EraCrypto era)) → Set (TxIn (EraCrypto era)) → StrictSeq (Sized (TxOut era)) → StrictMaybe (Sized (TxOut era)) → StrictMaybe Coin → StrictSeq (TxCert era) → Withdrawals (EraCrypto era) → Coin → ValidityInterval → StrictMaybe (Update era) → Set (KeyHash 'Witness (EraCrypto era)) → MultiAsset (EraCrypto era) → StrictMaybe (ScriptIntegrityHash (EraCrypto era)) → StrictMaybe (AuxiliaryDataHash (EraCrypto era)) → StrictMaybe Network → BabbageTxBody era

data BabbageTxBodyRaw era

data BabbageTxBodyUpgradeError

Constructors

BTBUEUpdatesD
The update attempts to update the decentralistion parameter, which is dropped in Babbage.

BTBUEUpdatesExtraEntropy
The update attempts to update the extra entropy, which is dropped in Babbage.

babbageAllInputsTxBodyF ∷ BabbageEraTxBody era ⇒ SimpleGetter (TxBody era) (Set (TxIn (EraCrypto era)))

babbageSpendableInputsTxBodyF ∷ BabbageEraTxBody era ⇒ SimpleGetter (TxBody era) (Set (TxIn (EraCrypto era)))

class (AlonzoEraTxBody era, BabbageEraTxOut era) ⇒ BabbageEraTxBody era where

Methods

sizedOutputsTxBodyL ∷ Lens' (TxBody era) (StrictSeq (Sized (TxOut era)))

referenceInputsTxBodyL ∷ Lens' (TxBody era) (Set (TxIn (EraCrypto era)))

totalCollateralTxBodyL ∷ Lens' (TxBody era) (StrictMaybe Coin)

collateralReturnTxBodyL ∷ Lens' (TxBody era) (StrictMaybe (TxOut era))

sizedCollateralReturnTxBodyL ∷ Lens' (TxBody era) (StrictMaybe (Sized (TxOut era)))

allSizedOutputsTxBodyF ∷ SimpleGetter (TxBody era) (StrictSeq (Sized (TxOut era)))

spendInputs' ∷ BabbageTxBody era → Set (TxIn (EraCrypto era))

collateralInputs' ∷ BabbageTxBody era → Set (TxIn (EraCrypto era))

referenceInputs' ∷ BabbageTxBody era → Set (TxIn (EraCrypto era))

outputs' ∷ BabbageTxBody era → StrictSeq (TxOut era)

collateralReturn' ∷ BabbageTxBody era → StrictMaybe (TxOut era)

totalCollateral' ∷ BabbageTxBody era → StrictMaybe Coin

certs' ∷ BabbageTxBody era → StrictSeq (TxCert era)

withdrawals' ∷ BabbageTxBody era → Withdrawals (EraCrypto era)

txfee' ∷ BabbageTxBody era → Coin

vldt' ∷ BabbageTxBody era → ValidityInterval

update' ∷ BabbageTxBody era → StrictMaybe (Update era)

reqSignerHashes' ∷ BabbageTxBody era → Set (KeyHash 'Witness (EraCrypto era))

mint' ∷ BabbageTxBody era → MultiAsset (EraCrypto era)

scriptIntegrityHash' ∷ BabbageTxBody era → StrictMaybe (ScriptIntegrityHash (EraCrypto era))

adHash' ∷ BabbageTxBody era → StrictMaybe (AuxiliaryDataHash (EraCrypto era))

txnetworkid' ∷ BabbageTxBody era → StrictMaybe Network

getEitherAddrBabbageTxOut ∷ (HasCallStack, HashAlgorithm (ADDRHASH (EraCrypto era))) ⇒ BabbageTxOut era → Either (Addr (EraCrypto era)) (CompactAddr (EraCrypto era))

data EraIndependentScriptIntegrity

type ScriptIntegrityHash c = SafeHash c EraIndependentScriptIntegrity

txOutData ∷ Era era ⇒ BabbageTxOut era → Maybe (Data era)

Deprecated: In favor of dataTxOutL or getDataBabbageTxOut

txOutDataHash ∷ Era era ⇒ BabbageTxOut era → Maybe (DataHash (EraCrypto era))

Deprecated: In favor of dataHashTxOutL or getDataHashBabbageTxOut

txOutScript ∷ BabbageTxOut era → Maybe (Script era)

Deprecated: In favor of dataTxOutL or getScriptBabbageTxOut

Orphan instances

Crypto c ⇒ AllegraEraTxBody (BabbageEra c)

Instance details

Methods

vldtTxBodyL ∷ Lens' (TxBody (BabbageEra c)) ValidityInterval

Crypto c ⇒ AlonzoEraTxBody (BabbageEra c)

Instance details

Methods

collateralInputsTxBodyL ∷ Lens' (TxBody (BabbageEra c)) (Set (TxIn (EraCrypto (BabbageEra c))))

reqSignerHashesTxBodyL ∷ Lens' (TxBody (BabbageEra c)) (Set (KeyHash 'Witness (EraCrypto (BabbageEra c))))

scriptIntegrityHashTxBodyL ∷ Lens' (TxBody (BabbageEra c)) (StrictMaybe (ScriptIntegrityHash (EraCrypto (BabbageEra c))))

networkIdTxBodyL ∷ Lens' (TxBody (BabbageEra c)) (StrictMaybe Network)

redeemerPointer ∷ TxBody (BabbageEra c) → PlutusPurpose AsItem (BabbageEra c) → StrictMaybe (PlutusPurpose AsIx (BabbageEra c))

redeemerPointerInverse ∷ TxBody (BabbageEra c) → PlutusPurpose AsIx (BabbageEra c) → StrictMaybe (PlutusPurpose AsIxItem (BabbageEra c))

Crypto c ⇒ EraTxBody (BabbageEra c)

Instance details

Associated Types

type TxBody (BabbageEra c) = (r ∷ Type)

type TxBodyUpgradeError (BabbageEra c)

Methods

mkBasicTxBody ∷ TxBody (BabbageEra c)

inputsTxBodyL ∷ Lens' (TxBody (BabbageEra c)) (Set (TxIn (EraCrypto (BabbageEra c))))

outputsTxBodyL ∷ Lens' (TxBody (BabbageEra c)) (StrictSeq (TxOut (BabbageEra c)))

feeTxBodyL ∷ Lens' (TxBody (BabbageEra c)) Coin

withdrawalsTxBodyL ∷ Lens' (TxBody (BabbageEra c)) (Withdrawals (EraCrypto (BabbageEra c)))

auxDataHashTxBodyL ∷ Lens' (TxBody (BabbageEra c)) (StrictMaybe (AuxiliaryDataHash (EraCrypto (BabbageEra c))))

spendableInputsTxBodyF ∷ SimpleGetter (TxBody (BabbageEra c)) (Set (TxIn (EraCrypto (BabbageEra c))))

allInputsTxBodyF ∷ SimpleGetter (TxBody (BabbageEra c)) (Set (TxIn (EraCrypto (BabbageEra c))))

certsTxBodyL ∷ Lens' (TxBody (BabbageEra c)) (StrictSeq (TxCert (BabbageEra c)))

getTotalDepositsTxBody ∷ PParams (BabbageEra c) → (KeyHash 'StakePool (EraCrypto (BabbageEra c)) → Bool) → TxBody (BabbageEra c) → Coin

getTotalRefundsTxBody ∷ PParams (BabbageEra c) → (Credential 'Staking (EraCrypto (BabbageEra c)) → Maybe Coin) → (Credential 'DRepRole (EraCrypto (BabbageEra c)) → Maybe Coin) → TxBody (BabbageEra c) → Coin

getGenesisKeyHashCountTxBody ∷ TxBody (BabbageEra c) → Int

upgradeTxBody ∷ TxBody (PreviousEra (BabbageEra c)) → Either (TxBodyUpgradeError (BabbageEra c)) (TxBody (BabbageEra c))

Crypto c ⇒ MaryEraTxBody (BabbageEra c)

Instance details

Methods

mintTxBodyL ∷ Lens' (TxBody (BabbageEra c)) (MultiAsset (EraCrypto (BabbageEra c)))

mintValueTxBodyF ∷ SimpleGetter (TxBody (BabbageEra c)) (Value (BabbageEra c))

mintedTxBodyF ∷ SimpleGetter (TxBody (BabbageEra c)) (Set (PolicyID (EraCrypto (BabbageEra c))))

Crypto c ⇒ ShelleyEraTxBody (BabbageEra c)

Instance details

Methods

ttlTxBodyL ∷ Lens' (TxBody (BabbageEra c)) SlotNo

updateTxBodyL ∷ Lens' (TxBody (BabbageEra c)) (StrictMaybe (Update (BabbageEra c)))