Cardano.Ledger.Babbage.TxBody.Internal

Description

Provides Babbage TxBody internals

Warning

This module is considered internal.

The contents of this module may change in any way whatsoever and without any warning between minor versions of this package.

Documentation

data BabbageTxOut era

Constructors

TxOutCompactDatum !CompactAddr !(CompactForm (Value era)) !(BinaryData era)

TxOutCompactRefScript !CompactAddr !(CompactForm (Value era)) !(Datum era) !(Script era)

Bundled Patterns

pattern BabbageTxOut ∷ (Era era, Val (Value era), HasCallStack) ⇒ Addr → Value era → Datum era → StrictMaybe (Script era) → BabbageTxOut era

pattern TxOutCompact ∷ (Era era, Val (Value era), Compactible (Value era), HasCallStack) ⇒ CompactAddr → CompactForm (Value era) → BabbageTxOut era

pattern TxOutCompactDH ∷ (Era era, Val (Value era), Compactible (Value era), HasCallStack) ⇒ CompactAddr → CompactForm (Value era) → DataHash → BabbageTxOut era

allSizedOutputsBabbageTxBodyF ∷ BabbageEraTxBody era ⇒ SimpleGetter (TxBody era) (StrictSeq (Sized (TxOut era)))

babbageMinUTxOValue ∷ BabbageEraPParams era ⇒ PParams era → Sized a → Coin

newtype BabbageTxBody era

Constructors

TxBodyConstr (MemoBytes (BabbageTxBodyRaw era))

Bundled Patterns

pattern BabbageTxBody ∷ ∀ era. BabbageEraTxBody era ⇒ Set TxIn → Set TxIn → Set TxIn → StrictSeq (Sized (TxOut era)) → StrictMaybe (Sized (TxOut era)) → StrictMaybe Coin → StrictSeq (TxCert era) → Withdrawals → Coin → ValidityInterval → StrictMaybe (Update era) → Set (KeyHash 'Witness) → MultiAsset → StrictMaybe ScriptIntegrityHash → StrictMaybe TxAuxDataHash → StrictMaybe Network → BabbageTxBody era

data BabbageTxBodyRaw era

Constructors

BabbageTxBodyRaw

Fields

btbrSpendInputs ∷ !(Set TxIn)

btbrCollateralInputs ∷ !(Set TxIn)

btbrReferenceInputs ∷ !(Set TxIn)

btbrOutputs ∷ !(StrictSeq (Sized (TxOut era)))

btbrCollateralReturn ∷ !(StrictMaybe (Sized (TxOut era)))

btbrTotalCollateral ∷ !(StrictMaybe Coin)

btbrCerts ∷ !(StrictSeq (TxCert era))

btbrWithdrawals ∷ !Withdrawals

btbrTxFee ∷ !Coin

btbrValidityInterval ∷ !ValidityInterval

btbrUpdate ∷ !(StrictMaybe (Update era))

btbrReqSignerHashes ∷ !(Set (KeyHash 'Witness))

btbrMint ∷ !MultiAsset

btbrScriptIntegrityHash ∷ !(StrictMaybe ScriptIntegrityHash)

btbrAuxDataHash ∷ !(StrictMaybe TxAuxDataHash)

btbrTxNetworkId ∷ !(StrictMaybe Network)

data BabbageTxBodyUpgradeError

Constructors

BTBUEUpdatesD
The update attempts to update the decentralistion parameter, which is dropped in Babbage.

BTBUEUpdatesExtraEntropy
The update attempts to update the extra entropy, which is dropped in Babbage.

babbageAllInputsTxBodyF ∷ BabbageEraTxBody era ⇒ SimpleGetter (TxBody era) (Set TxIn)

babbageSpendableInputsTxBodyF ∷ BabbageEraTxBody era ⇒ SimpleGetter (TxBody era) (Set TxIn)

class (AlonzoEraTxBody era, BabbageEraTxOut era) ⇒ BabbageEraTxBody era where

Methods

sizedOutputsTxBodyL ∷ Lens' (TxBody era) (StrictSeq (Sized (TxOut era)))

referenceInputsTxBodyL ∷ Lens' (TxBody era) (Set TxIn)

totalCollateralTxBodyL ∷ Lens' (TxBody era) (StrictMaybe Coin)

collateralReturnTxBodyL ∷ Lens' (TxBody era) (StrictMaybe (TxOut era))

sizedCollateralReturnTxBodyL ∷ Lens' (TxBody era) (StrictMaybe (Sized (TxOut era)))

allSizedOutputsTxBodyF ∷ SimpleGetter (TxBody era) (StrictSeq (Sized (TxOut era)))

spendInputs' ∷ BabbageTxBody era → Set TxIn

collateralInputs' ∷ BabbageTxBody era → Set TxIn

referenceInputs' ∷ BabbageTxBody era → Set TxIn

outputs' ∷ BabbageTxBody era → StrictSeq (TxOut era)

collateralReturn' ∷ BabbageTxBody era → StrictMaybe (TxOut era)

totalCollateral' ∷ BabbageTxBody era → StrictMaybe Coin

certs' ∷ BabbageTxBody era → StrictSeq (TxCert era)

withdrawals' ∷ BabbageTxBody era → Withdrawals

txfee' ∷ BabbageTxBody era → Coin

vldt' ∷ BabbageTxBody era → ValidityInterval

update' ∷ BabbageTxBody era → StrictMaybe (Update era)

reqSignerHashes' ∷ BabbageTxBody era → Set (KeyHash 'Witness)

mint' ∷ BabbageTxBody era → MultiAsset

scriptIntegrityHash' ∷ BabbageTxBody era → StrictMaybe ScriptIntegrityHash

adHash' ∷ BabbageTxBody era → StrictMaybe TxAuxDataHash

txnetworkid' ∷ BabbageTxBody era → StrictMaybe Network

getEitherAddrBabbageTxOut ∷ BabbageTxOut era → Either Addr CompactAddr

data EraIndependentScriptIntegrity

type ScriptIntegrityHash = SafeHash EraIndependentScriptIntegrity

txOutData ∷ Era era ⇒ BabbageTxOut era → Maybe (Data era)

Deprecated: In favor of dataTxOutL or getDataBabbageTxOut

txOutDataHash ∷ BabbageTxOut era → Maybe DataHash

Deprecated: In favor of dataHashTxOutL or getDataHashBabbageTxOut

txOutScript ∷ BabbageTxOut era → Maybe (Script era)

Deprecated: In favor of dataTxOutL or getScriptBabbageTxOut

Orphan instances

AllegraEraTxBody BabbageEra

Instance details

Methods

vldtTxBodyL ∷ Lens' (TxBody BabbageEra) ValidityInterval

AlonzoEraTxBody BabbageEra

Instance details

Methods

collateralInputsTxBodyL ∷ Lens' (TxBody BabbageEra) (Set TxIn)

reqSignerHashesTxBodyL ∷ Lens' (TxBody BabbageEra) (Set (KeyHash 'Witness))

scriptIntegrityHashTxBodyL ∷ Lens' (TxBody BabbageEra) (StrictMaybe ScriptIntegrityHash)

networkIdTxBodyL ∷ Lens' (TxBody BabbageEra) (StrictMaybe Network)

redeemerPointer ∷ TxBody BabbageEra → PlutusPurpose AsItem BabbageEra → StrictMaybe (PlutusPurpose AsIx BabbageEra)

redeemerPointerInverse ∷ TxBody BabbageEra → PlutusPurpose AsIx BabbageEra → StrictMaybe (PlutusPurpose AsIxItem BabbageEra)

EraTxBody BabbageEra

Instance details

Associated Types

type TxBody BabbageEra = (r ∷ Type)

type TxBodyUpgradeError BabbageEra

Methods

mkBasicTxBody ∷ TxBody BabbageEra

inputsTxBodyL ∷ Lens' (TxBody BabbageEra) (Set TxIn)

outputsTxBodyL ∷ Lens' (TxBody BabbageEra) (StrictSeq (TxOut BabbageEra))

feeTxBodyL ∷ Lens' (TxBody BabbageEra) Coin

withdrawalsTxBodyL ∷ Lens' (TxBody BabbageEra) Withdrawals

auxDataHashTxBodyL ∷ Lens' (TxBody BabbageEra) (StrictMaybe TxAuxDataHash)

spendableInputsTxBodyF ∷ SimpleGetter (TxBody BabbageEra) (Set TxIn)

allInputsTxBodyF ∷ SimpleGetter (TxBody BabbageEra) (Set TxIn)

certsTxBodyL ∷ Lens' (TxBody BabbageEra) (StrictSeq (TxCert BabbageEra))

getTotalDepositsTxBody ∷ PParams BabbageEra → (KeyHash 'StakePool → Bool) → TxBody BabbageEra → Coin

getTotalRefundsTxBody ∷ PParams BabbageEra → (Credential 'Staking → Maybe Coin) → (Credential 'DRepRole → Maybe Coin) → TxBody BabbageEra → Coin

getGenesisKeyHashCountTxBody ∷ TxBody BabbageEra → Int

upgradeTxBody ∷ TxBody (PreviousEra BabbageEra) → Either (TxBodyUpgradeError BabbageEra) (TxBody BabbageEra)

MaryEraTxBody BabbageEra

Instance details

Methods

mintTxBodyL ∷ Lens' (TxBody BabbageEra) MultiAsset

mintValueTxBodyF ∷ SimpleGetter (TxBody BabbageEra) (Value BabbageEra)

mintedTxBodyF ∷ SimpleGetter (TxBody BabbageEra) (Set PolicyID)

ShelleyEraTxBody BabbageEra

Instance details

Methods

ttlTxBodyL ∷ Lens' (TxBody BabbageEra) SlotNo

updateTxBodyL ∷ Lens' (TxBody BabbageEra) (StrictMaybe (Update BabbageEra))