Cardano.Ledger.Babbage.TxWits

Documentation

upgradeTxDats ∷ (Era era1, Era era2, EraCrypto era1 ~ EraCrypto era2) ⇒ TxDats era1 → TxDats era2

Upgrade TxDats from one era to another. The underlying data structure will remain identical, but the memoised serialisation may change to reflect the versioned serialisation of the new era.

upgradeRedeemers ∷ (AlonzoEraScript (PreviousEra era), AlonzoEraScript era) ⇒ Redeemers (PreviousEra era) → Redeemers era

Upgrade redeemers from one era to another. The underlying data structure will remain identical, but the memoised serialisation may change to reflect the versioned serialisation of the new era.

data AlonzoTxWits era where

Bundled Patterns

pattern AlonzoTxWits ∷ AlonzoEraScript era ⇒ Set (WitVKey 'Witness (EraCrypto era)) → Set (BootstrapWitness (EraCrypto era)) → Map (ScriptHash (EraCrypto era)) (Script era) → TxDats era → Redeemers era → AlonzoTxWits era

pattern AlonzoTxWits' ∷ Era era ⇒ Set (WitVKey 'Witness (EraCrypto era)) → Set (BootstrapWitness (EraCrypto era)) → Map (ScriptHash (EraCrypto era)) (Script era) → TxDats era → Redeemers era → AlonzoTxWits era

class (EraTxWits era, AlonzoEraScript era) ⇒ AlonzoEraTxWits era where

Methods

datsTxWitsL ∷ Lens' (TxWits era) (TxDats era)

rdmrsTxWitsL ∷ Lens' (TxWits era) (Redeemers era)

Orphan instances

Crypto c ⇒ AlonzoEraTxWits (BabbageEra c)

Instance details

Methods

datsTxWitsL ∷ Lens' (TxWits (BabbageEra c)) (TxDats (BabbageEra c))

rdmrsTxWitsL ∷ Lens' (TxWits (BabbageEra c)) (Redeemers (BabbageEra c))

Crypto c ⇒ EraTxWits (BabbageEra c)

Instance details

Associated Types

type TxWits (BabbageEra c) = (r ∷ Type)

Methods

mkBasicTxWits ∷ TxWits (BabbageEra c)

addrTxWitsL ∷ Lens' (TxWits (BabbageEra c)) (Set (WitVKey 'Witness (EraCrypto (BabbageEra c))))

bootAddrTxWitsL ∷ Lens' (TxWits (BabbageEra c)) (Set (BootstrapWitness (EraCrypto (BabbageEra c))))

scriptTxWitsL ∷ Lens' (TxWits (BabbageEra c)) (Map (ScriptHash (EraCrypto (BabbageEra c))) (Script (BabbageEra c)))

upgradeTxWits ∷ TxWits (PreviousEra (BabbageEra c)) → TxWits (BabbageEra c)