Cardano.Ledger.Babbage.UTxO

Documentation

getBabbageSupplementalDataHashes ∷ BabbageEraTxBody era ⇒ UTxO era → TxBody era → Set (DataHash (EraCrypto era))

getBabbageSpendingDatum ∷ (AlonzoEraTx era, BabbageEraTxOut era) ⇒ UTxO era → Tx era → PlutusPurpose AsItem era → Maybe (Data era)

Extract binary data either directly from the Tx as an "inline datum" or look it up in the witnesses by the hash.

getBabbageScriptsProvided ∷ (EraTx era, BabbageEraTxBody era) ⇒ UTxO era → Tx era → ScriptsProvided era

getReferenceScripts ∷ BabbageEraTxOut era ⇒ UTxO era → Set (TxIn (EraCrypto era)) → Map (ScriptHash (EraCrypto era)) (Script era)

Collect all the reference scripts found in the TxOuts, pointed to by some input.

getReferenceScriptsNonDistinct ∷ BabbageEraTxOut era ⇒ UTxO era → Set (TxIn (EraCrypto era)) → [(ScriptHash (EraCrypto era), Script era)]

Orphan instances

Crypto c ⇒ AlonzoEraUTxO (BabbageEra c)

Instance details

Methods

getSupplementalDataHashes ∷ UTxO (BabbageEra c) → TxBody (BabbageEra c) → Set (DataHash (EraCrypto (BabbageEra c)))

getSpendingDatum ∷ UTxO (BabbageEra c) → Tx (BabbageEra c) → PlutusPurpose AsItem (BabbageEra c) → Maybe (Data (BabbageEra c))

Crypto c ⇒ EraUTxO (BabbageEra c)

Instance details

Associated Types

type ScriptsNeeded (BabbageEra c) = (r ∷ Type)

Methods

getConsumedValue ∷ PParams (BabbageEra c) → (Credential 'Staking (EraCrypto (BabbageEra c)) → Maybe Coin) → (Credential 'DRepRole (EraCrypto (BabbageEra c)) → Maybe Coin) → UTxO (BabbageEra c) → TxBody (BabbageEra c) → Value (BabbageEra c)

getProducedValue ∷ PParams (BabbageEra c) → (KeyHash 'StakePool (EraCrypto (BabbageEra c)) → Bool) → TxBody (BabbageEra c) → Value (BabbageEra c)

getScriptsProvided ∷ UTxO (BabbageEra c) → Tx (BabbageEra c) → ScriptsProvided (BabbageEra c)

getScriptsNeeded ∷ UTxO (BabbageEra c) → TxBody (BabbageEra c) → ScriptsNeeded (BabbageEra c)

getScriptsHashesNeeded ∷ ScriptsNeeded (BabbageEra c) → Set (ScriptHash (EraCrypto (BabbageEra c)))

getWitsVKeyNeeded ∷ CertState (BabbageEra c) → UTxO (BabbageEra c) → TxBody (BabbageEra c) → Set (KeyHash 'Witness (EraCrypto (BabbageEra c)))

getMinFeeTxUtxo ∷ PParams (BabbageEra c) → Tx (BabbageEra c) → UTxO (BabbageEra c) → Coin