Cardano.Ledger.Babbage.PParams

Description

This module contains just the type of protocol parameters.

Documentation

data PParams' f era

Protocol parameters. Alonzo parameters without d and extraEntropy

Constructors

PParams

Fields

_minfeeA :: !(HKD f Natural)

The linear factor for the minimum fee calculation

_minfeeB :: !(HKD f Natural)

The constant factor for the minimum fee calculation

_maxBBSize :: !(HKD f Natural)

Maximal block body size

_maxTxSize :: !(HKD f Natural)

Maximal transaction size

_maxBHSize :: !(HKD f Natural)

Maximal block header size

_keyDeposit :: !(HKD f Coin)

The amount of a key registration deposit

_poolDeposit :: !(HKD f Coin)

The amount of a pool registration deposit

_eMax :: !(HKD f EpochNo)

Maximum number of epochs in the future a pool retirement is allowed to be scheduled for.

_nOpt :: !(HKD f Natural)

Desired number of pools

_a0 :: !(HKD f NonNegativeInterval)

Pool influence

_rho :: !(HKD f UnitInterval)

Monetary expansion

_tau :: !(HKD f UnitInterval)

Treasury expansion

_protocolVersion :: !(HKD f ProtVer)

Protocol version

_minPoolCost :: !(HKD f Coin)

Minimum Stake Pool Cost

_coinsPerUTxOByte :: !(HKD f Coin)

Cost in lovelace per byte of UTxO storage (instead of _coinsPerUTxOByte)

_costmdls :: !(HKD f CostModels)

Cost models for non-native script languages

_prices :: !(HKD f Prices)

Prices of execution units (for non-native script languages)

_maxTxExUnits :: !(HKD f ExUnits)

Max total script execution resources units allowed per tx

_maxBlockExUnits :: !(HKD f ExUnits)

Max total script execution resources units allowed per block

_maxValSize :: !(HKD f Natural)

Max size of a Value in an output

_collateralPercentage :: !(HKD f Natural)

Percentage of the txfee which must be provided as collateral when including non-native scripts.

_maxCollateralInputs :: !(HKD f Natural)

Maximum number of collateral inputs allowed in a transaction

type PParams = PParams' Identity

emptyPParams :: PParams era

Returns a basic "empty" PParams structure with all zero values.

type PParamsUpdate era = PParams' StrictMaybe era

emptyPParamsUpdate :: PParamsUpdate era

updatePParams :: PParams era -> PParamsUpdate era -> PParams era

Update operation for protocol parameters structure @PParams

getLanguageView :: HasField "_costmdls" (PParams era) CostModels => PParams era -> Language -> LangDepView

data LangDepView

Constructors

LangDepView

Fields

tag :: ByteString

params :: ByteString

encodeLangViews :: Set LangDepView -> Encoding

retractPP :: HKD f Coin -> HKD f UnitInterval -> HKD f Nonce -> PParams' f era -> PParams' f era

Turn an PParams' into a Shelley.Params'

extendPP :: PParams' f era1 -> HKD f Coin -> HKD f CostModels -> HKD f Prices -> HKD f ExUnits -> HKD f ExUnits -> HKD f Natural -> HKD f Natural -> HKD f Natural -> PParams' f era2

Given the missing pieces Turn a Shelley.PParams' into an Params'