Cardano.Ledger.Babbage.PParams

Description

This module contains the type of protocol parameters and EraPParams instance

Documentation

data BabbagePParams f era

Babbage Protocol parameters. Ways in which parameters have changed from Alonzo: lack of d, extraEntropy and replacement of coinsPerUTxOWord with coinsPerUTxOByte

Constructors

BabbagePParams

Fields

bppMinFeeA ∷ !(HKD f Coin)

The linear factor for the minimum fee calculation

bppMinFeeB ∷ !(HKD f Coin)

The constant factor for the minimum fee calculation

bppMaxBBSize ∷ !(HKD f Natural)

Maximal block body size

bppMaxTxSize ∷ !(HKD f Natural)

Maximal transaction size

bppMaxBHSize ∷ !(HKD f Natural)

Maximal block header size

bppKeyDeposit ∷ !(HKD f Coin)

The amount of a key registration deposit

bppPoolDeposit ∷ !(HKD f Coin)

The amount of a pool registration deposit

bppEMax ∷ !(HKD f EpochNo)

Maximum number of epochs in the future a pool retirement is allowed to be scheduled for.

bppNOpt ∷ !(HKD f Natural)

Desired number of pools

bppA0 ∷ !(HKD f NonNegativeInterval)

Pool influence

bppRho ∷ !(HKD f UnitInterval)

Monetary expansion

bppTau ∷ !(HKD f UnitInterval)

Treasury expansion

bppProtocolVersion ∷ !(HKD f ProtVer)

Protocol version

bppMinPoolCost ∷ !(HKD f Coin)

Minimum Stake Pool Cost

bppCoinsPerUTxOByte ∷ !(HKD f CoinPerByte)

Cost in lovelace per byte of UTxO storage (instead of bppCoinsPerUTxOByte)

bppCostModels ∷ !(HKD f CostModels)

Cost models for non-native script languages

bppPrices ∷ !(HKD f Prices)

Prices of execution units (for non-native script languages)

bppMaxTxExUnits ∷ !(HKD f OrdExUnits)

Max total script execution resources units allowed per tx

bppMaxBlockExUnits ∷ !(HKD f OrdExUnits)

Max total script execution resources units allowed per block

bppMaxValSize ∷ !(HKD f Natural)

Max size of a Value in an output

bppCollateralPercentage ∷ !(HKD f Natural)

Percentage of the txfee which must be provided as collateral when including non-native scripts.

bppMaxCollateralInputs ∷ !(HKD f Natural)

Maximum number of collateral inputs allowed in a transaction

emptyBabbagePParams ∷ ∀ era. Era era ⇒ BabbagePParams Identity era

Returns a basic "empty" PParams structure with all zero values.

emptyBabbagePParamsUpdate ∷ BabbagePParams StrictMaybe era

data DowngradeBabbagePParams f

Constructors

DowngradeBabbagePParams

Fields

dbppD ∷ !(HKD f UnitInterval)

dbppExtraEntropy ∷ !(HKD f Nonce)

getLanguageView ∷ AlonzoEraPParams era ⇒ PParams era → Language → LangDepView

data LangDepView

Constructors

LangDepView

Fields

tag ∷ ByteString

params ∷ ByteString

encodeLangViews ∷ Set LangDepView → Encoding

coinsPerUTxOWordToCoinsPerUTxOByte ∷ CoinPerWord → CoinPerByte

A word is 8 bytes, so convert from coinsPerUTxOWord to coinsPerUTxOByte, rounding down.

coinsPerUTxOByteToCoinsPerUTxOWord ∷ CoinPerByte → CoinPerWord

A word is 8 bytes, so convert from coinsPerUTxOByte to coinsPerUTxOWord.

Orphan instances

Crypto c ⇒ AlonzoEraPParams (BabbageEra c)

Instance details

Methods

hkdCoinsPerUTxOWordL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, ExactEra AlonzoEra (BabbageEra c)) ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f CoinPerWord)

hkdCostModelsL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f CostModels)

hkdPricesL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f Prices)

hkdMaxTxExUnitsL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f ExUnits)

hkdMaxBlockExUnitsL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f ExUnits)

hkdMaxValSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f Natural)

hkdCollateralPercentageL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f Natural)

hkdMaxCollateralInputsL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f Natural)

Crypto c ⇒ BabbageEraPParams (BabbageEra c)

Instance details

Methods

hkdCoinsPerUTxOByteL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f CoinPerByte)

Crypto c ⇒ EraPParams (BabbageEra c)

Instance details

Associated Types

type PParamsHKD f (BabbageEra c) = (r ∷ Type)

type UpgradePParams f (BabbageEra c)

type DowngradePParams f (BabbageEra c)

Methods

applyPPUpdates ∷ PParams (BabbageEra c) → PParamsUpdate (BabbageEra c) → PParams (BabbageEra c)

emptyPParamsIdentity ∷ PParamsHKD Identity (BabbageEra c)

emptyPParamsStrictMaybe ∷ PParamsHKD StrictMaybe (BabbageEra c)

upgradePParamsHKD ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, EraPParams (PreviousEra (BabbageEra c))) ⇒ UpgradePParams f (BabbageEra c) → PParamsHKD f (PreviousEra (BabbageEra c)) → PParamsHKD f (BabbageEra c)

downgradePParamsHKD ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, EraPParams (PreviousEra (BabbageEra c))) ⇒ DowngradePParams f (BabbageEra c) → PParamsHKD f (BabbageEra c) → PParamsHKD f (PreviousEra (BabbageEra c))

hkdMinFeeAL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f Coin)

hkdMinFeeBL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f Coin)

hkdMaxBBSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f Natural)

hkdMaxTxSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f Natural)

hkdMaxBHSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f Natural)

hkdKeyDepositL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f Coin)

hkdPoolDepositL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f Coin)

hkdEMaxL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f EpochNo)

hkdNOptL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f Natural)

hkdA0L ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f NonNegativeInterval)

hkdRhoL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f UnitInterval)

hkdTauL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f UnitInterval)

hkdDL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, ProtVerAtMost (BabbageEra c) 6) ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f UnitInterval)

ppDG ∷ SimpleGetter (PParams (BabbageEra c)) UnitInterval

hkdExtraEntropyL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, ProtVerAtMost (BabbageEra c) 6) ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f Nonce)

hkdProtocolVersionL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f ProtVer)

hkdMinUTxOValueL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, ProtVerAtMost (BabbageEra c) 4) ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f Coin)

hkdMinPoolCostL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (BabbageEra c)) (HKD f Coin)

Crypto c ⇒ EraGovernance (BabbageEra c)

Instance details

Associated Types

type GovernanceState (BabbageEra c) = (r ∷ Type)

Methods

emptyGovernanceState ∷ GovernanceState (BabbageEra c)

getProposedPPUpdates ∷ GovernanceState (BabbageEra c) → Maybe (ProposedPPUpdates (BabbageEra c))