Cardano.Ledger.Babbage.TxBody

Documentation

data TxOut era

Constructors

TxOutCompactDatum !(CompactAddr (Crypto era)) !(CompactForm (Value era)) !(BinaryData era)

TxOutCompactRefScript !(CompactAddr (Crypto era)) !(CompactForm (Value era)) !(Datum era) !(Script era)

Bundled Patterns

pattern TxOut :: forall era. (Era era, Compactible (Value era), Val (Value era), HasCallStack) => Addr (Crypto era) -> Value era -> Datum era -> StrictMaybe (Script era) -> TxOut era

pattern TxOutCompact :: (Era era, HasCallStack) => CompactAddr (Crypto era) -> CompactForm (Value era) -> TxOut era

pattern TxOutCompactDH :: (Era era, HasCallStack) => CompactAddr (Crypto era) -> CompactForm (Value era) -> DataHash (Crypto era) -> TxOut era

data TxBody era where

Bundled Patterns

pattern TxBody :: BabbageBody era => Set (TxIn (Crypto era)) -> Set (TxIn (Crypto era)) -> Set (TxIn (Crypto era)) -> StrictSeq (Sized (TxOut era)) -> StrictMaybe (Sized (TxOut era)) -> StrictMaybe Coin -> StrictSeq (DCert (Crypto era)) -> Wdrl (Crypto era) -> Coin -> ValidityInterval -> StrictMaybe (Update era) -> Set (KeyHash 'Witness (Crypto era)) -> Value (Crypto era) -> StrictMaybe (ScriptIntegrityHash (Crypto era)) -> StrictMaybe (AuxiliaryDataHash (Crypto era)) -> StrictMaybe Network -> TxBody era

data Datum era

Datum can be described by a either a data hash or binary data, but not both. It can also be neither one of them.

Constructors

NoDatum

DatumHash !(DataHash (Crypto era))

Datum !(BinaryData era)

spendInputs' :: TxBody era -> Set (TxIn (Crypto era))

collateralInputs' :: TxBody era -> Set (TxIn (Crypto era))

referenceInputs' :: TxBody era -> Set (TxIn (Crypto era))

outputs' :: TxBody era -> StrictSeq (TxOut era)

collateralReturn' :: TxBody era -> StrictMaybe (TxOut era)

totalCollateral' :: TxBody era -> StrictMaybe Coin

certs' :: TxBody era -> StrictSeq (DCert (Crypto era))

wdrls' :: TxBody era -> Wdrl (Crypto era)

txfee' :: TxBody era -> Coin

vldt' :: TxBody era -> ValidityInterval

update' :: TxBody era -> StrictMaybe (Update era)

reqSignerHashes' :: TxBody era -> Set (KeyHash 'Witness (Crypto era))

mint' :: TxBody era -> Value (Crypto era)

scriptIntegrityHash' :: TxBody era -> StrictMaybe (ScriptIntegrityHash (Crypto era))

adHash' :: TxBody era -> StrictMaybe (AuxiliaryDataHash (Crypto era))

txnetworkid' :: TxBody era -> StrictMaybe Network

getBabbageTxOutEitherAddr :: HashAlgorithm (ADDRHASH (Crypto era)) => TxOut era -> Either (Addr (Crypto era)) (CompactAddr (Crypto era))

type BabbageBody era = (Era era, ToCBOR (Value era), ToCBOR (Script era), SerialisableData (PParamsDelta era))

data EraIndependentScriptIntegrity

type ScriptIntegrityHash crypto = SafeHash crypto EraIndependentScriptIntegrity

txOutData :: TxOut era -> Maybe (Data era)

txOutDataHash :: Era era => TxOut era -> Maybe (DataHash (Crypto era))

Return the data hash of a given transaction output, if one is present. Note that this function does *not* return the hash of any inline datums that are present.

txOutScript :: TxOut era -> Maybe (Script era)