Cardano.Ledger.Babbage.TxInfo

Documentation

data BabbageContextError era

Constructors

AlonzoContextError !(AlonzoContextError era)

ByronTxOutInContext !(TxOutSource (EraCrypto era))

RedeemerPointerPointsToNothing !(PlutusPurpose AsIx era)

InlineDatumsNotSupported !(TxOutSource (EraCrypto era))

ReferenceScriptsNotSupported !(TxOutSource (EraCrypto era))

ReferenceInputsNotSupported !(Set (TxIn (EraCrypto era)))

transReferenceScript ∷ ∀ era. EraScript era ⇒ StrictMaybe (Script era) → Maybe ScriptHash

transTxOutV1 ∷ ∀ era. (Inject (BabbageContextError era) (ContextError era), Value era ~ MaryValue (EraCrypto era), BabbageEraTxOut era) ⇒ TxOutSource (EraCrypto era) → TxOut era → Either (ContextError era) TxOut

Given a TxOut, translate it for V2 and return (Right transalation). If the transaction contains any Byron addresses or Babbage features, return Left.

transTxOutV2 ∷ ∀ era. (Inject (BabbageContextError era) (ContextError era), Value era ~ MaryValue (EraCrypto era), BabbageEraTxOut era) ⇒ TxOutSource (EraCrypto era) → TxOut era → Either (ContextError era) TxOut

Given a TxOut, translate it for V2 and return (Right transalation). It is possible the address part is a Bootstrap Address, in that case return Left.

transTxInInfoV1 ∷ ∀ era. (Inject (BabbageContextError era) (ContextError era), Value era ~ MaryValue (EraCrypto era), BabbageEraTxOut era) ⇒ UTxO era → TxIn (EraCrypto era) → Either (ContextError era) TxInInfo

Given a TxIn, look it up in the UTxO. If it exists, translate it to the V1 context

transTxInInfoV2 ∷ ∀ era. (Inject (BabbageContextError era) (ContextError era), Value era ~ MaryValue (EraCrypto era), BabbageEraTxOut era) ⇒ UTxO era → TxIn (EraCrypto era) → Either (ContextError era) TxInInfo

Given a TxIn, look it up in the UTxO. If it exists, translate it to the V2 context

transTxRedeemers ∷ (EraPlutusTxInfo l era, AlonzoEraTxBody era, EraTx era, AlonzoEraTxWits era, Inject (BabbageContextError era) (ContextError era)) ⇒ proxy l → ProtVer → Tx era → Either (ContextError era) (Map (PlutusScriptPurpose l) Redeemer)

Translate all Redeemers from within a Tx into a Map from a PlutusScriptPurpose to a Redeemer

transRedeemer ∷ Data era → Redeemer

toPlutusV2Args ∷ EraPlutusTxInfo 'PlutusV2 era ⇒ proxy 'PlutusV2 → ProtVer → TxInfo → PlutusPurpose AsIxItem era → Maybe (Data era) → Data era → Either (ContextError era) (PlutusArgs 'PlutusV2)

Orphan instances

Crypto c ⇒ EraPlutusTxInfo 'PlutusV1 (BabbageEra c)

Instance details

Methods

toPlutusTxCert ∷ proxy 'PlutusV1 → ProtVer → TxCert (BabbageEra c) → Either (ContextError (BabbageEra c)) (PlutusTxCert 'PlutusV1)

toPlutusScriptPurpose ∷ proxy 'PlutusV1 → ProtVer → PlutusPurpose AsIxItem (BabbageEra c) → Either (ContextError (BabbageEra c)) (PlutusScriptPurpose 'PlutusV1)

toPlutusTxInfo ∷ proxy 'PlutusV1 → LedgerTxInfo (BabbageEra c) → Either (ContextError (BabbageEra c)) (PlutusTxInfo 'PlutusV1)

toPlutusArgs ∷ proxy 'PlutusV1 → ProtVer → PlutusTxInfo 'PlutusV1 → PlutusPurpose AsIxItem (BabbageEra c) → Maybe (Data (BabbageEra c)) → Data (BabbageEra c) → Either (ContextError (BabbageEra c)) (PlutusArgs 'PlutusV1)

Crypto c ⇒ EraPlutusTxInfo 'PlutusV2 (BabbageEra c)

Instance details

Methods

toPlutusTxCert ∷ proxy 'PlutusV2 → ProtVer → TxCert (BabbageEra c) → Either (ContextError (BabbageEra c)) (PlutusTxCert 'PlutusV2)

toPlutusScriptPurpose ∷ proxy 'PlutusV2 → ProtVer → PlutusPurpose AsIxItem (BabbageEra c) → Either (ContextError (BabbageEra c)) (PlutusScriptPurpose 'PlutusV2)

toPlutusTxInfo ∷ proxy 'PlutusV2 → LedgerTxInfo (BabbageEra c) → Either (ContextError (BabbageEra c)) (PlutusTxInfo 'PlutusV2)

toPlutusArgs ∷ proxy 'PlutusV2 → ProtVer → PlutusTxInfo 'PlutusV2 → PlutusPurpose AsIxItem (BabbageEra c) → Maybe (Data (BabbageEra c)) → Data (BabbageEra c) → Either (ContextError (BabbageEra c)) (PlutusArgs 'PlutusV2)

Crypto c ⇒ EraPlutusContext (BabbageEra c)

Instance details

Associated Types

type ContextError (BabbageEra c) = (r ∷ Type)

Methods

mkPlutusWithContext ∷ PlutusScript (BabbageEra c) → ScriptHash (EraCrypto (BabbageEra c)) → PlutusPurpose AsIxItem (BabbageEra c) → LedgerTxInfo (BabbageEra c) → (Data (BabbageEra c), ExUnits) → CostModel → Either (ContextError (BabbageEra c)) (PlutusWithContext (EraCrypto (BabbageEra c)))