Cardano.Ledger.Babbage.TxInfo

Documentation

transScriptHash ∷ ScriptHash c → ScriptHash

transReferenceScript ∷ ∀ era. EraScript era ⇒ StrictMaybe (Script era) → Maybe ScriptHash

txInfoOutV1 ∷ ∀ era. (BabbageEraTxOut era, Value era ~ MaryValue (EraCrypto era)) ⇒ TxOutSource (EraCrypto era) → TxOut era → Either (TranslationError (EraCrypto era)) TxOut

Given a TxOut, translate it for V2 and return (Right transalation). If the transaction contains any Byron addresses or Babbage features, return Left.

txInfoOutV2 ∷ ∀ era. (BabbageEraTxOut era, Value era ~ MaryValue (EraCrypto era)) ⇒ TxOutSource (EraCrypto era) → TxOut era → Either (TranslationError (EraCrypto era)) TxOut

Given a TxOut, translate it for V2 and return (Right transalation). It is possible the address part is a Bootstrap Address, in that case return Left.

txInfoInV1 ∷ ∀ era. (BabbageEraTxOut era, Value era ~ MaryValue (EraCrypto era)) ⇒ UTxO era → TxIn (EraCrypto era) → Either (TranslationError (EraCrypto era)) TxInInfo

Given a TxIn, look it up in the UTxO. If it exists, translate it to the V1 context and return (Just translation). If does not exist in the UTxO, return Nothing.

txInfoInV2 ∷ ∀ era. (BabbageEraTxOut era, Value era ~ MaryValue (EraCrypto era)) ⇒ UTxO era → TxIn (EraCrypto era) → Either (TranslationError (EraCrypto era)) TxInInfo

Given a TxIn, look it up in the UTxO. If it exists, translate it to the V2 context and return (Just translation). If does not exist in the UTxO, return Nothing.

transRedeemer ∷ Data era → Redeemer

transRedeemerPtr ∷ (MaryEraTxBody era, EraPlutusContext 'PlutusV1 era) ⇒ TxBody era → (RdmrPtr, (Data era, ExUnits)) → Either (TranslationError (EraCrypto era)) (ScriptPurpose, Redeemer)

babbageTxInfo ∷ ∀ era. (EraTx era, AlonzoEraTxWits era, BabbageEraTxBody era, Value era ~ MaryValue (EraCrypto era), EraPlutusContext 'PlutusV1 era, EraPlutusContext 'PlutusV2 era) ⇒ PParams era → Language → EpochInfo (Either Text) → SystemStart → UTxO era → Tx era → Either (TranslationError (EraCrypto era)) VersionedTxInfo

babbageTxInfoV1 ∷ ∀ era. (EraTx era, AlonzoEraTxWits era, BabbageEraTxBody era, Value era ~ MaryValue (EraCrypto era), EraPlutusContext 'PlutusV1 era) ⇒ POSIXTimeRange → Tx era → UTxO era → Either (TranslationError (EraCrypto era)) VersionedTxInfo

babbageTxInfoV2 ∷ ∀ era. (EraTx era, AlonzoEraTxWits era, BabbageEraTxBody era, Value era ~ MaryValue (EraCrypto era), EraPlutusContext 'PlutusV2 era, EraPlutusContext 'PlutusV1 era) ⇒ POSIXTimeRange → Tx era → UTxO era → Either (TranslationError (EraCrypto era)) VersionedTxInfo

Orphan instances

Crypto c ⇒ EraPlutusContext 'PlutusV1 (BabbageEra c)

Instance details

Methods

transTxCert ∷ TxCert (BabbageEra c) → PlutusTxCert 'PlutusV1

Crypto c ⇒ EraPlutusContext 'PlutusV2 (BabbageEra c)

Instance details

Methods

transTxCert ∷ TxCert (BabbageEra c) → PlutusTxCert 'PlutusV2